0.9.2.3: fix multiple idiot errors from importing

+++ Extensions/Item.cs
@@ -113,13 +113,14 @@ public partial class Item
 
     [XmlIgnore] public Gender Gender => Properties.Restrictions?.Gender ?? Gender.Neutral;
 
-    [XmlIgnore] public ushort MinLDamage => Properties.Damage?.Large.Min ?? 0;
+    [XmlIgnore]
+    public float MinLDamage => Properties.Damage?.Large.Min ?? 0;
 
-    [XmlIgnore] public ushort MaxLDamage => Properties.Damage?.Large.Max ?? 0;
+    [XmlIgnore] public float MaxLDamage => Properties.Damage?.Large.Max ?? 0;
 
-    [XmlIgnore] public ushort MinSDamage => Properties.Damage?.Small.Min ?? 0;
+    [XmlIgnore] public float MinSDamage => Properties.Damage?.Small.Min ?? 0;
 
-    [XmlIgnore] public ushort MaxSDamage => Properties.Damage?.Small.Max ?? 0;
+    [XmlIgnore] public float MaxSDamage => Properties.Damage?.Small.Max ?? 0;
 
     [XmlIgnore] public Variant CurrentVariant { get; set; }
